backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
PyJWT==2.8.0
bcrypt==4.1.1
httpx==0.25.1

# AI/ML
openai==1.3.0
google-generativeai==0.3.0
faster-whisper==0.10.0
numpy==1.24.3
scipy==1.11.3

# PDF
reportlab==4.0.7
PyPDF2==3.0.1

# Database
alembic==1.12.1
sqlalchemy-utils==0.41.1

# Cloud
supabase==1.0.3
boto3==1.28.82

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-jose==3.3.0
email-validator==2.1.0
